Activate cart price

+++ app/src/main/java/com/morj12/cloudlist/domain/entity/Cart.kt
@@ -2,5 +2,5 @@ package com.morj12.cloudlist.domain.entity
 
 data class Cart(
     val timestamp: Long = 0,
-    val price: Long = 0
+    val price: Double = 0.0
 )
